requirements.txt
streamlit
opencv-python-headless
ultralytics
pillow